changes to ORBMatcher to take l2-norm and threshold values. They need to be fine tuned

+++ include/ORBmatcher.h
@@ -90,8 +90,8 @@ namespace ORB_SLAM3
 
     public:
 
-        static const int TH_LOW;
-        static const int TH_HIGH;
+        static const double TH_LOW;
+        static const double TH_HIGH;
         static const int HISTO_LENGTH;
         EIGEN_MAKE_ALIGNED_OPERATOR_NEW
